Update file saving and texture filtering behavior
Added std::ios_base::trunc to all file output streams to ensure files are properly truncated on save. Changed texture filtering to use GL_NEAREST for both direct and compressed textures to match in-game MBAACC behavior. Fixed WriteAF to write color data as int instead of float. Simplified logic for selecting the save target HA6 file in ini.cpp. Updated imgui submodule.

+++ src/framedata_save.cpp
@@ -74,7 +74,7 @@ void WriteAF(std::ofstream &file, const Frame_AF *af)
 				for(int j = 0; j < 3; j++)
 					anormalized[j] = layer.rgba[j]*255.f;
 				file.write("AFRG", 4);
-				file.write(PTR(anormalized), 3*sizeof(float));
+				file.write(PTR(anormalized), 3*sizeof(int));
 			}
 
 			if(layer.rotation[0]){
@@ -129,7 +129,7 @@ void WriteAF(std::ofstream &file, const Frame_AF *af)
 			for(int i = 0; i < 3; i++)
 				anormalized[i] = layer.rgba[i]*255.f;
 			file.write("AFRG", 4);
-			file.write(PTR(anormalized), 3*sizeof(float));
+			file.write(PTR(anormalized), 3*sizeof(int));
 		}
 
 		if(layer.rotation[0]){
